Fixes from previous version: - changed some naming for main - splitted win32 functions from all the rest for later refactoring
