+++ ProtX/source/protx/protx/ProtX.cpp
@@ -4,25 +4,11 @@
 
 #include "protx/protx/ProtX.h"
 
-#include <Windows.h>
-#include "protx/window/PWindow.h"
-
 protx::ProtX::ProtX() {
     m_display = std::unique_ptr<PWindow>(new PWindow());
 }
 
 protx::ProtX::~ProtX() {
-}
-
-void protx::ProtX::run() {
-    MSG msg;
-    while (m_isRunning && !m_display->isClosed()) {
-        if (PeekMessage(&msg, NULL, NULL, NULL, PM_REMOVE)) {
-            TranslateMessage(&msg);
-            DispatchMessage(&msg);
-        }
-        Sleep(1);
-    }
 }
 
 void protx::ProtX::quit() {
